Add stricter checking when decoding with .from_payload()
Earlier, it was possible that a protobuf payload not including a certain
message was decoded using .from_payload(). For example when using
GetConfigsRequest.from_payload() with a message missing
GenericMessage::wirepas::get_configs_req, the library was not reporting
an error and setting the req_id to 0 even though it was not initialized
in the protobuf message.

If that happens now, and InvalidMessageType exception is raised. This
exception has the GatewayAPIParsingException as its parent to help with
backwards compatibility.

Also refactoring the tests and having a new test file for decoding
related error testing.